+++ eva-accession-import-dbsnp2/src/main/java/uk/ac/ebi/eva/accession/dbsnp2/processors/ContigToGenbankReplacerProcessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.dbsnp2.io;
+package uk.ac.ebi.eva.accession.dbsnp2.processors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
